# openapi_schema.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "CosyVoice API",
    "description": "Cross-lingual Voice Cloning API with CosyVoice2 (v2) and CosyVoice3 (v3)",
    "version": "3.0.0",
    "contact": {
      "name": "CosyVoice API Support"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8012",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/api/v3/cross-lingual/with-audio": {
      "post": {
        "tags": ["Cross-lingual Synthesis (v3)"],
        "summary": "Synthesize with audio reference",
        "operationId": "cross_lingual_with_audio_v3",
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CrossLingualWithAudioRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Synthesis result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SynthesisResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/cross-lingual/with-cache": {
      "post": {
        "tags": ["Cross-lingual Synthesis (v3)"],
        "summary": "Synthesize with cached voice",
        "operationId": "cross_lingual_with_cache_v3",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrossLingualWithCacheRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Synthesis result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SynthesisResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v3/cross-lingual/task": {
      "post": {
        "tags": ["Cross-lingual Synthesis (v3)"],
        "summary": "Create synthesis task",
        "operationId": "create_task_v3",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v3/cross-lingual/task/{task_id}": {
      "get": {
        "tags": ["Cross-lingual Synthesis (v3)"],
        "summary": "Get task status",
        "operationId": "get_task_status_v3",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskStatusResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Cross-lingual Synthesis (v3)"],
        "summary": "Delete task",
        "operationId": "delete_task_v3",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Task deleted"
          }
        }
      }
    },
...
    "/api/v3/audio/{filename}": {
      "get": {
        "tags": ["Cross-lingual Synthesis (v3)"],
        "summary": "Download audio file",
        "operationId": "get_audio_v3",
        "parameters": [
          {
            "name": "filename",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Audio file",
            "content": {
              "audio/wav": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AudioFormat": {
        "type": "string",
        "enum": ["wav", "mp3", "flac"],
        "default": "wav"
      },
...
      "CrossLingualWithAudioRequest": {
        "type": "object",
        "required": ["text", "prompt_audio"],
        "properties": {
          "text": {
            "type": "string",
            "description": "Text to synthesize",
            "maxLength": 2000
          },
          "format": {
            "$ref": "#/components/schemas/AudioFormat"
          },
          "speed": {
            "type": "number",
            "default": 1.0,
            "minimum": 0.5,
            "maximum": 2.0
          },
          "stream": {
            "type": "boolean",
            "default": false
          },
          "instruct_text": {
            "type": "string",
            "description": "V3 only: Instruction for voice control"
          },
          "prompt_audio": {
            "type": "string",
            "format": "binary",
            "description": "Reference audio file"
          }
        }
      },
      "CrossLingualWithCacheRequest": {
        "type": "object",
        "required": ["text", "voice_id"],
        "properties": {
          "text": {
            "type": "string",
            "description": "Text to synthesize",
            "maxLength": 2000
          },
          "voice_id": {
            "type": "string",
            "description": "Cached voice ID"
          },
          "format": {
            "$ref": "#/components/schemas/AudioFormat"
          },
          "speed": {
            "type": "number",
            "default": 1.0,
            "minimum": 0.5,
            "maximum": 2.0
          },
          "stream": {
            "type": "boolean",
            "default": false
          }
        }
      },
...
      "SynthesisResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "audio_url": {
            "type": "string",
            "description": "URL to download audio"
          },
          "file_path": {
            "type": "string"
          },
          "duration": {
            "type": "number",
            "description": "Audio duration in seconds"
          },
          "format": {
            "$ref": "#/components/schemas/AudioFormat"
          },
          "synthesis_time": {
            "type": "number",
            "description": "Processing time in seconds"
          }
        }
      },
      "TaskRequest": {
        "type": "object",
        "required": ["text", "voice_id"],
        "properties": {
          "text": {
            "type": "string",
            "maxLength": 1000
          },
          "voice_id": {
            "type": "string"
          },
          "format": {
            "type": "string",
            "default": "wav"
          },
          "speed": {
            "type": "number",
            "default": 1.0,
            "minimum": 0.5,
            "maximum": 2.0
          },
          "instruct_text": {
            "type": "string"
          }
        }
      },
      "TaskResponse": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "string",
            "format": "uuid"
          },
          "status": {
            "type": "string",
            "enum": ["registered", "processing", "completed", "failed"]
          },
          "file_path": {
            "type": "string"
          },
          "audio_url": {
            "type": "string"
          },
          "estimated_duration": {
            "type": "number"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "TaskStatusResponse": {
        "type": "object",
        "properties": {
          "task_id": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": ["registered", "processing", "completed", "failed"]
          },
          "file_path": {
            "type": "string"
          },
          "audio_url": {
            "type": "string"
          },
          "progress": {
            "type": "number",
            "minimum": 0,
            "maximum": 1
          },
          "duration": {
            "type": "number"
          },
          "synthesis_time": {
            "type": "number"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "completed_at": {
            "type": "string",
            "format": "date-time"
          },
          "error_message": {
            "type": "string"
          }
        }
      },
...
    }
  }
}